[CVS-10651]: Added binding for licenseUsageImpl and register license resouce

+++ 300-cv-nextgen/src/main/java/io/harness/cvng/core/services/api/monitoredService/MonitoredServiceService.java
@@ -108,5 +108,5 @@ public interface MonitoredServiceService extends DeleteEntityByHandler<Monitored
   PageResponse<NotificationRuleResponse> getNotificationRules(
       ProjectParams projectParams, String monitoredServiceIdentifier, PageParams pageParams);
   void beforeNotificationRuleDelete(ProjectParams projectParams, String notificationRuleRef);
-  int countUniqueEnabledServices(String accountId);
+  long countUniqueEnabledServices(String accountId);
 }
